Tidy up before pushing to production
